+++ providers/ViewProvider.js
@@ -28,6 +28,7 @@ class ViewProvider extends ServiceProvider {
       const View = require('../src/View')
       return new View(Helpers, Config.get('app.views.cache'))
     })
+    this.app.alias('Adonis/Src/View', 'View')
   }
 
   /**